projectFinish

+++ client/src/components/Card/utils/handleDelete.js
@@ -1,13 +1,34 @@
 import axios from "axios";
 import { filterDeleted } from "../../../Redux/actions";
-export const handleDelete = async (id,dispatch) => {
-    try {
-      dispatch(filterDeleted(id))
-      const {data} = await axios.delete(
+import Swal from "sweetalert2";
+
+export const handleDelete = async (id, dispatch) => {
+  try {
+    const result = await Swal.fire({
+      title: "Are you sure to remove this pokemon?",
+      text: "This action can not be undone!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete!",
+      cancelButtonText: "Cancel",
+    });
+
+    if (result.isConfirmed) {
+      dispatch(filterDeleted(id));
+      const { data } = await axios.delete(
         `http://localhost:3001/pokemons/deletePokemon/${id}`
       );
-      console.log(data)
-    } catch (error) {
-      console.error(error);
+      console.log(data);
+
+      Swal.fire(
+        "Pokemon has been deleted",
+        "The pokemon has been removed successfully",
+        "success"
+      );
     }
-  };
+  } catch (error) {
+    console.error(error);
+  }
+};
